refactor(ui): 动画引擎优化 — 速度边界裁剪、返回值统一、多维度settled判定修复
修复:
- A1: xerintosh_animation() 增加speed下限裁剪(ANIM_SPEED_MIN)
- A2: 魔数提取为 ANIM_SPEED_MAX/MIN/SNAP_THRESHOLD 常量
- A3: xerintosh_animation() 返回bool指示稳定状态
- A4: anim_row 多维度(y+w) settled判定修复
- A5: scroll_offset 纳入 settled 判定
- A7: 退场动画屏幕检测类型显式转换

优化:
- A10: 弹簧动画Euler积分特征值稳定性证明注释

测试:
- 新增7个 AnimationTest 用例(吸附/边界/禁用/返回值/负方向)

验证: hw构建通过, native 224测试通过, 0新增警告

+++ src/ui/ui_draw_anim.c
@@ -131,10 +131,10 @@ void xerintosh_draw_exit_animation()
 
     /* 屏幕方向/尺寸切换时，防止目标值残留旧尺寸 */
     if (g_xerintosh_exit_anim_prev_screen_h != SCREEN_HEIGHT) {
-        float max_h = SCREEN_HEIGHT + 8;
+        float max_h = (float)SCREEN_HEIGHT + 8.0f;
         if (g_xerintosh_exit_anim_temp_h > max_h) g_xerintosh_exit_anim_temp_h = max_h;
         if (g_xerintosh_exit_animation_status == 0) g_xerintosh_exit_anim_temp_h_trg = max_h;
-        g_xerintosh_exit_anim_prev_screen_h = SCREEN_HEIGHT;
+        g_xerintosh_exit_anim_prev_screen_h = (int16_t)SCREEN_HEIGHT;
     }
 
     /* 绘制全屏黑色遮罩 */
